requirements.txt
python-telegram-bot==20.7
aiohttp==3.9.3
apscheduler==3.10.4
anthropic==0.21.3
openai==1.14.3
